requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
audio-utils==0.0.1
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
coverage==5.5
cryptography==3.4.8
cupshelpers==1.0
Cython==3.0.11
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
eyed3==0.9.7
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flatbuffers==24.12.23
fsspec==2024.12.0
gradio==5.11.0
gradio_client==1.5.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.8.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
linkify-it-py==2.0.3
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
moondream==0.0.6
more-itertools==8.10.0
mpmath==1.3.0
mutagen==1.47.0
netifaces==0.11.0
numpy==2.2.0
oauthlib==3.2.0
olefile==0.46
onnx==1.17.0
onnxruntime==1.20.1
openai==1.59.5
opencv-python==4.10.0.84
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.2
ptyprocess==0.7.0
PyAudio==0.2.11
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzed @ file:///tmp/selfgz129160/pyzed-4.2-cp310-cp310-linux_x86_64.whl
reportlab==3.6.8
requests==2.25.1
rich==13.9.4
ruff==0.8.6
safehttpx==0.1.6
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
ssh-import-id==5.11
starlette==0.41.3
sympy==1.13.3
systemd-python==234
textual==1.0.0
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.3
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uv==0.5.16
uvicorn==0.34.0
wadllib==1.3.6
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
whisper==1.1.10
xdg==5
xkit==0.0.0
zipp==1.0.0
